WIP: pretty status general framework
The current proposed framework for including the human-readable status
is to create a new `PrettyBootEntry` struct that returns a subset of the
BootEntry fields. Both the original `BootEntry` and new `PrettyBootEntry`
structs will be returned from the `boot_entry_from_deployment()`
function and a conditional argument will be used to determine which
of the two to use. We've opted to create a new Struct since it provides
more flexiblity if we decide to change what fields the human-readable
status includes. It also allows us to have the conditional check for
the `pretty` flag to be outside the `boot_entry_from_deployment()`
function.

Co-authored-by: Joseph Marrero <jmarrero@redhat.com>
Co-authored-by: Huijing Hei <hhei@redhat.com>
Co-authored-by: Yasmin de Souza <ydesouza@redhat.com>
Co-authored-by: Steven Presti <spresti@redhat.com>

Signed-off-by: Luke Yang <luyang@redhat.com>