install: Detect aboot images and enable the correct bootloader
Aboot images need to have the ostree bootloader backend set
to "aboot", otherwise the deploy during bootc install will
not create the correct boot A/B symlinks, and additionally
once booted will not correctly deploy to the aboot partition
during an update.

To see whethere an image is using aboot, we look for the "aboot.img"
file in the kernel modules dir.

NOTE: In order to correctly handle running bootc from a different
container than the to-be-installed container we look for aboot.img
in the actual commit, not just in the running container.